Move btree tests to own file
This change moves the btree integration tests to their own file.  This
should improve readability of the submodules.
